Add serialization to beverages
Beverages will be rendered as objects with a root key of "beverage"
or "beverages" that is an array of objects representing the beverage and
its recipe. The beverage objects have a key, "name", with its name as
the value and a key, "recipe", that is an array of objects representing
the ingredients. The ingredients object has a key, "name", with its name
as the value, and a key, "parts", with the amount of relative parts to
be used.