Add serialization to ingredients
Ingredients will be rendered as objects with a root key of "ingredient"
or "ingredients" that is an array of objects representing the ingredient
and the beverages that it is used in. The ingredient objects have a key,
"name", with its name as the value and a key, "beverages", that is an
array of strings representing the beverages.